mmc: MMC card capacity calculation fix.
MMC card capacity calculation was truncated to 32 bits.
This resulted in incorrect values being reported
for large MMC cards.


Review URL: http://codereview.chromium.org/2035014
Patch from Amol Jadi <ajadi@quicinc.com>.

+++ drivers/mmc/mmc.c
@@ -702,7 +702,7 @@ int mmc_startup(struct mmc *mmc)
 	u64 cmult, csize;
 	struct mmc_cmd cmd;
 	char ext_csd[512];
-	u32 sec_count;
+	u64 sec_count;
 
 	/* Put the Card in Identify Mode */
 	cmd.cmdidx = MMC_CMD_ALL_SEND_CID;
